feat: implement getWithAutCookie utility for enhanced cookie handling
